- 아이콘 출력 오류 수정 - 이미지 리사이즈 오류 수정

+++ apk-extractor/app/src/main/java/net/sarangnamu/apk_extractor/Common.kt
@@ -49,8 +49,8 @@ fun Drawable.bitmap(config: Bitmap.Config = Bitmap.Config.ARGB_8888): Bitmap {
 
 fun Bitmap.resize(w: Int, h: Int): Bitmap {
     val matrix = Matrix()
-    val scaleW = (w / width).toFloat()
-    val scaleH = (h / height).toFloat()
+    val scaleW = w.toFloat() / width.toFloat()
+    val scaleH = h.toFloat() / height.toFloat()
 
     matrix.postScale(scaleW, scaleH)
